nl2postcondition-fse2024/nl2postcondition_source_evalplus/requirements.txt
wget
appdirs
tempdir
multipledispatch
rich
openai==0.27.0
aiohttp==3.9.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==4.2.2
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pyzmq==25.0.0
requests==2.28.2
six==1.16.0
stack-data==0.6.2
tenacity==8.2.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
yarl==1.8.2
zipp==3.15.0
python-decouple==3.8
evalplus==0.1.5
hydra-core==1.3.2
docker==6.1.3
